Highlight the text within paragraphs rather than paragraphs themselves
The TED example highlights text within paragraphs with whitespace
between vertical lines. If we highlight the entire paragraph it will
also highlight this whitespace.

Therefore, introduce a ‘mark’ child element that wraps the contents of
the paragraph and highlight that instead. Set a line-height of 1.5rem so
we can see that it does the right thing.